ecommerce_flask/requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
supabase==2.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
requests-oauthlib==1.3.1